State Management. Save Progress #4. Book List Refactor Done.
