+++ src/app/pages/books-list/core/domain/book-paged-data.ts
