+++ src/app/pages/books-list/core/domain/book-summary.ts
