+++ src/app/pages/books-list/core/domain/index.ts
