+++ src/app/pages/books-list/core/index.ts
@@ -1,3 +1,4 @@
 export * from './actions';
 export * from './state';
+export * from './domain';
 export { BookListFacade } from './book-list.facade';
